Allows non folder exports (#1026)
I was unable to export the dataset from UI, when I only specified the
name of the file.
The problem. is that python can't create empty directories, but if we
don't specify one the code still attempts to create a directory. This
will cause an error.

This PR fixes it